+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where a Neo batch actions bar's paste button wasn't being disabled if pasting the copied blocks would violate the Neo field's Max Blocks or Max Top Level Blocks settings
+
 ## 1.2.1 - 2023-02-21
 
 ### Added